refactor: 계산 결과 탭에서 활용을 위한 필터링된 적금 상품 전역 상태로 이동
